[FIX] point_of_sale: preserve unpaid orders with a payment
Unpaid orders are preserved in the local storage and can be recovered
even if you close the session and open a new one. But they would be
removed if they didn't contain any lines (articles).

On the other hand, it's possible to remove all of the order's lines
after receiving payment. So it's essential to preserve an order which
contains a payment.

Steps to reproduce the issue:
	1_ Open a PoS session and create a new order
	2_ Add some articles and proceed with the payment
	3_ Select one of the payment methods to make the payment
	4_ Do not validate the session and click on the "Back" button
	5_ Remove all of the order's article and close the session
	6_ Open a new session from the same PoS
=> The order with the received payment is removed

The solution is to preserve the orders with a payment line.

opw-2766658

closes odoo/odoo#87223

X-original-commit: 3445786415371612b2fc7075223e201c3882e0c6
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>
Signed-off-by: Masereel Pierre <pim@odoo.com>